general: linux audio extract now considers actual header len
The documentation on page /general/audio-file-format/ section 2. Audio file
format reports, that the header is of variable length. Thus, for precise
audio data extraction, the header's length must be parsed in order to
discard the correct number of bytes.

Previously, the code-snippet discarded always 4096 bytes, which would fail
when the header length should be ever different.
Also, i always like to have easy-to-copy&paste snippets in documentation.